user_service/serializer.py
from user_service.models import Question, Choice
from rest_framework.serializers import ModelSerializer


class QuestionSerializer(ModelSerializer):
    class Meta:
        model = Question
        fields = '__all__'


class ChoiceSerializer(ModelSerializer):
    class Meta:
        model = Choice
        fields = '__all__'
