requirements.txt
django
djangorestframework
django-admin
django-rest-swagger
gunicorn
pygments
psycopg2-binary
drf-yasg
drf-yasg[validation]